(create) set size/stroke-color actions/cases
