+++ client/store/actions/types.js
@@ -5,6 +5,8 @@ export const GAME_TYPES = {
   SET_IS_ADMIN: 'SET_IS_ADMIN',
   SET_CONTEXT: 'SET_CONTEXT',
   SET_TOOL: 'SET_TOOL',
+  SET_SIZE: 'SET_SIZE',
+  SET_STROKE_COLOR: 'SET_STROKE_COLOR',
   SET_FILL: 'SET_FILL',
   SET_FILL_COLOR: 'SET_FILL_COLOR',
